refactor: Migrate command implementations from main.go to commands package
Major architectural cleanup to improve code organization and maintainability:

### Key Changes:
- **main.go**: Reduced from ~900+ lines to 245 lines (70% reduction)
- **Commands migrated to dedicated files**:
  - devcontainer → cmd/claude-reactor/commands/devcontainer.go
  - template → cmd/claude-reactor/commands/template.go
  - dependency → cmd/claude-reactor/commands/dependency.go
  - hotreload → cmd/claude-reactor/commands/hotreload.go
  - debug → cmd/claude-reactor/commands/debug.go
  - completion → cmd/claude-reactor/commands/completion.go

### Technical Improvements:
- **Consistent naming patterns**: All commands use NewXXXCmd(app) constructors
- **Proper separation of concerns**: Each command in its own file
- **Version handling**: Implemented SetVersionInfo() for debug command
- **Clean imports**: Removed unused imports (time, encoding/json, path/filepath)
- **Modular architecture**: Commands package provides clean interface

### Quality Assurance:
- ✅ All unit tests pass (100% success rate)
- ✅ All integration tests pass
- ✅ Zero functional regressions
- ✅ Clean compilation with no warnings
- ✅ Maintained backward compatibility

This refactoring creates a much cleaner, more maintainable codebase while
preserving all existing functionality. main.go is now focused purely on
application initialization and command registration.
